functions/openai_sync/requirements.txt
google-cloud-firestore
openai